update texdesigner warning

+++ Graphic11/textdesigner/DiffusedShadowStrategy.cpp
@@ -74,13 +74,13 @@ bool DiffusedShadowStrategy::DrawString(
 {
 	using namespace Gdiplus;
 	GraphicsPath path;
-	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,nfontSize,ptDraw,pStrFormat);
+	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,(REAL)nfontSize,ptDraw,pStrFormat);
 	if(status!=Ok)
 		return false;
 
 	for(int i=1; i<=m_nThickness; ++i)
 	{
-		Pen pen(m_clrOutline,i);
+		Pen pen(m_clrOutline, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, &path);
 	}
@@ -114,13 +114,13 @@ bool DiffusedShadowStrategy::DrawString(
 {
 	using namespace Gdiplus;
 	GraphicsPath path;
-	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,nfontSize,rtDraw,pStrFormat);
+	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,(REAL)nfontSize,rtDraw,pStrFormat);
 	if(status!=Ok)
 		return false;
 
 	for(int i=1; i<=m_nThickness; ++i)
 	{
-		Pen pen(m_clrOutline,i);
+		Pen pen(m_clrOutline, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, &path);
 	}
@@ -170,7 +170,7 @@ bool DiffusedShadowStrategy::GdiDrawString(
 
 	for(int i=1; i<=m_nThickness; ++i)
 	{
-		Pen pen(m_clrOutline,i);
+		Pen pen(m_clrOutline, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, pPath);
 	}
@@ -225,7 +225,7 @@ bool DiffusedShadowStrategy::GdiDrawString(
 
 	for(int i=1; i<=m_nThickness; ++i)
 	{
-		Pen pen(m_clrOutline,i);
+		Pen pen(m_clrOutline, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, pPath);
 	}
